+++ sdk/communication/Azure.Communication.CallAutomation/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1-alpha.7 (Unreleased)
  - DialogContext in StartDialogOption is now of type IDictionary<string, object>
  - StatDialog now takes dialogId of type Guid
+ - Remove TargetParticipant from StartDialogOptions
 
 ## 1.0.1-alpha.6 (Unreleased)
 - Added Conversation Conductor events